tbi-backend/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
google-auth
google-cloud-documentai
